requirements.txt
# Core NLP libraries
nltk>=3.8.1
spacy>=3.5.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Web framework
streamlit>=1.28.0

# API clients
google-generativeai>=0.3.0
requests>=2.31.0

# ML/AI libraries
torch>=2.0.0
transformers>=4.35.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0